requirements.txt
numpy~=1.19.4
matplotlib~=3.3.2
pillow~=8.0.1
pandas~=1.1.3
scikit-learn~=0.23.2
tensorflow~=2.3.0